requirements.txt
# --- Core API & Environment ---
openai>=1.0.0
python-dotenv>=1.0.0

# --- Local Intelligence (Offline STT) ---
faster-whisper>=1.0.0

# --- Audio Engine (System: portaudio19-dev python3-dev) ---
pyaudio>=0.2.14
webrtcvad>=2.0.10

# --- Compatibility & Utilities ---
setuptools>=69.0.0